+++ models/yolo3_model.py
@@ -11,7 +11,7 @@ from keras.layers.normalization import BatchNormalization
 from keras.models import Model
 from keras.regularizers import l2
 
-from yolo3.utils import compose
+from models.utils import compose
 
 
 @wraps(Conv2D)
