+++ yolo_detect/detect_image.py
@@ -1,8 +1,8 @@
 import numpy as np
-from yolo import YOLO
 from PIL import Image
 
-def detect_img(yolo):
+
+def detect_image(yolo):
     while True:
         try:
             input_img_path = input('Input image filename:')
@@ -17,13 +17,3 @@ def detect_img(yolo):
             output_img_path = input('Output image filename:')
             cv2.imwrite(output_img_path, np.array(r_image, dtype='float32'))
     yolo.close_session()
-
-FLAGS = None
-
-if __name__ == '__main__':
-
-    """
-    Image detection mode, disregard any remaining command line arguments
-    """
-    print("Image detection mode")
-    detect_img(YOLO())
